requirements.txt

click==8.1.7
Flask==3.0.3
gitdb==4.0.11
Jinja2==3.1.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
